Inline Client.Session
